RR-48 multi glyph auto (and some other things) (#19)
* RR-48 add opmode to test vision pose estimation

* RR-48 pathing and vision working pretty well independently

* RR-48 standardize timescales

* RR-48 add basic vision-based pose estimation

* RR-48 remove cryptobox vision localizer

* RR-48 update cryptobox position test opmode

* RR-48 add vision update to multi-glyph auto

* RR-48 added enable/disable functionality to trackers and fix multi-glyph auto

* RR-48 one step closer to multi-glyph auto; things kinda work

* RR-48 fix closest cryptobox determination

* RR-48 refactor camera logic

* RR-48 add OpenCV camera view support

* RR-48 add VuMark identification based on feature detection

* RR-48 fix overlay scaling issue

* RR-48 fixed some minor issues

* RR-48 small miscellaneous modifications

* RR-48 add hub benchmark

* RR-48 add intermediates for debugging vision

* RR-48 fix looper issues

* RR-48 add PID strafe segment to multi glyph auto

* RR-48 modifications to multiglyph auto

* RR-48 refactor cryptobox localization code into CryptoboxLocalizer

* RR-48 reorganize code in preparation for vision test app

* RR-48 add basic vision test app and fix crypto tracker point bug

* RR-48 add support for cycling through intermediate mats in VisionTest

* RR-48 fix issue with UI operations not executing on the UI thread

* RR-48 fix expansion hub benchmark and fix embedded IMU read window

* RR-48 add android platform 26 to .travis.yml

+++ TeamCode/src/main/java/com/acmerobotics/relicrecovery/loops/Loop.java
@@ -1,9 +1,12 @@
 package com.acmerobotics.relicrecovery.loops;
 
+import com.acmerobotics.library.util.TimestampedData;
+
 /**
  * Created by ryanbrott on 10/28/17.
  */
 
 public interface Loop {
-    void onLoop(long timestamp, long dt);
+    /** Uses the timescale described in {@link TimestampedData} */
+    void onLoop(double timestamp, double dt);
 }
